requirements.txt
# --------- core --------- #
numpy
scipy
Cython

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0

# --------- visualization --------- #
matplotlib

# --------- notebook --------- #
jupyter
ipywidgets

# --------- testing --------- #
pytest
pytest-cov
